\name{combinePrices}
\alias{combinePrices}
\title{Combine the Adjust close column of multiple xts objects}
\usage{
  combinePrices(symbols)
}
\arguments{
  \item{symbols}{character vector of symbols in your
  current workspace}
}
\value{
  xts object of adjusted closing prices
}
\description{
  This function combines the Adjusted closing prices of
  multiple xts objects.
}

